feat: improve abstraction of CRUD operations

+++ services/task.service.ts
@@ -1,5 +1,5 @@
 import { type DocumentDefinition } from "mongoose";
-import columnModel, { type IColumnModel } from "../models/column.model.js";
+import { type IColumnModel } from "../models/column.model.js";
 import { type Request, type Response } from "express";
 import { StatusCodes } from "http-status-codes";
 import taskModel, { type ITaskModel } from "../models/task.model.js";
@@ -11,17 +11,21 @@ import ETableName from "../constants/eTableName.js";
 import find from "../utils/databaseUtils/find.js";
 import createItem from "../utils/databaseUtils/create.js";
 import findByIdAndUpdate from "../utils/databaseUtils/findByIdAndUpdate.js";
+import ITask from "../interfaces/task.js";
+import IColumn from "../interfaces/column.js";
+import IUser from "../interfaces/user.js";
+import IProject from "../interfaces/project.js";
 
 const create = async (
     reqBody: DocumentDefinition<ITaskModel>,
     res: Response
 ): Promise<Response<any, Record<string, any>>> => {
     const { columnId, coordinatorId, projectId } = reqBody;
-    const column = await findById(columnId, ETableName.COLUMN);
-    const coordinator = await findById(coordinatorId, ETableName.USER);
-    const project = await findById(projectId, ETableName.PROJECT);
+    const column = await findById<IColumn>(columnId, ETableName.COLUMN);
+    const coordinator = await findById<IUser>(coordinatorId, ETableName.USER);
+    const project = await findById<IProject>(projectId, ETableName.PROJECT);
     if (column != null && coordinator != null && project != null) {
-        const tasks = await find({ columnId }, ETableName.TASK);
+        const tasks = await find<ITask>({ columnId }, ETableName.TASK);
         await createItem({ ...reqBody, index: tasks?.length }, ETableName.TASK);
         return res.status(StatusCodes.CREATED).json("Task created");
     } else {
@@ -36,40 +40,44 @@ const get = async (
     res: Response
 ): Promise<Response<any, Record<string, any>>> => {
     const { projectId } = req.query;
-    const columns = await find({ projectId }, ETableName.COLUMN);
-    if (columns != null && columns.length > 0) {
-        for (const c of columns as IColumnModel[]) {
-            const allTasks = await find({ projectId }, ETableName.TASK);
-            if (allTasks?.length === 0) {
-                if (c.columnName === "To Do") {
-                    await taskModel.create({
-                        columnId: c._id,
-                        projectId,
-                        taskName: "Default Task",
-                        coordinatorId: getUserId(req),
-                        epic: "Default epic",
-                        type: "Task",
-                        note: "No note yet",
-                        storyPoints: 1,
-                        index: 0
-                    });
+    if (typeof projectId === "string") {
+        const columns = await find<IColumn>({ projectId }, ETableName.COLUMN);
+        if (columns != null && columns.length > 0) {
+            for (const c of columns as IColumnModel[]) {
+                const allTasks = await find<ITask>({ projectId }, ETableName.TASK);
+                if (allTasks?.length === 0) {
+                    if (c.columnName === "To Do") {
+                        await createItem({
+                            columnId: c._id,
+                            projectId,
+                            taskName: "Default Task",
+                            coordinatorId: getUserId(req),
+                            epic: "Default epic",
+                            type: "Task",
+                            note: "No note yet",
+                            storyPoints: 1,
+                            index: 0
+                        }, ETableName.TASK);
+                    }
                 }
             }
+            const tasks = await find<ITask>({ projectId }, ETableName.TASK);
+            quickSort(tasks as Array<{ index: number }>);
+            return res.status(StatusCodes.OK).json(tasks);
+        } else {
+            return res.status(StatusCodes.NOT_FOUND).json("Column not found");
         }
-        const tasks = await find({ projectId }, ETableName.TASK);
-        quickSort(tasks as Array<{ index: number }>);
-        return res.status(StatusCodes.OK).json(tasks);
     } else {
-        return res.status(StatusCodes.NOT_FOUND).json("Column not found");
+        return res.status(StatusCodes.BAD_REQUEST).json("Bad request");
     }
 };
 
 const update = async (
-    reqBody: DocumentDefinition<ITaskModel>,
+    reqBody: ITask & { _id: string },
     res: Response
 ): Promise<Response<any, Record<string, any>>> => {
     const taskId = reqBody._id;
-    const task = await findById(taskId, ETableName.TASK);
+    const task = await findById<ITask>(taskId, ETableName.TASK);
     if (task != null) {
         await findByIdAndUpdate(taskId, reqBody, ETableName.TASK);
         return res.status(StatusCodes.OK).json("Task updated");
@@ -83,68 +91,72 @@ const remove = async (
     res: Response
 ): Promise<Response<any, Record<string, any>>> => {
     const { taskId } = req.query;
-    const task = await taskModel.findById(taskId);
-    if (task != null) {
-        await taskModel.findByIdAndDelete(taskId);
-        return res.status(StatusCodes.OK).json("Task deleted");
+    if (typeof taskId === "string") {
+        const task = await findById<ITask>(taskId, ETableName.TASK);
+        if (task != null) {
+            await taskModel.findByIdAndDelete(taskId);
+            return res.status(StatusCodes.OK).json("Task deleted");
+        } else {
+            return res.status(StatusCodes.NOT_FOUND).json("Task not found");
+        }
     } else {
-        return res.status(StatusCodes.NOT_FOUND).json("Task not found");
+        return res.status(StatusCodes.BAD_REQUEST).json("Bad request");
     }
 };
 
 const reorder = async (
-    reqBody: DocumentDefinition<ITaskOrder>,
+    reqBody: ITaskOrder,
     res: Response
 ): Promise<Response<any, Record<string, any>> | undefined> => {
     const { type, fromId, referenceId, fromColumnId, referenceColumnId } =
         reqBody;
-    const fromColumn = await columnModel.findById(fromColumnId);
-    const referenceColumn = await columnModel.findById(referenceColumnId);
-    const fromTask = await taskModel.findById(fromId);
-    const referenceTask = await taskModel.findById(referenceId);
+    const fromColumn = await findById<IColumn>(fromColumnId, ETableName.COLUMN);
+    const referenceColumn = await findById<IColumn>(referenceColumnId, ETableName.COLUMN);
+    const fromTask = await findById<ITask>(fromId, ETableName.TASK);
+    const referenceTask = await findById<ITask>(referenceId, ETableName.TASK);
     if (
         fromColumn != null &&
         referenceColumn != null &&
         fromTask != null &&
         (referenceId == null || referenceTask != null)
     ) {
-        const fromColumnTasks = await taskModel.find({
+        const fromColumnTasks = await find<ITask>({
             columnId: fromColumnId
-        });
-        const referenceColumnTasks = await taskModel.find({
+        }, ETableName.TASK);
+        const referenceColumnTasks = await find<ITask>({
             columnId: referenceColumnId
-        });
-        if (fromColumnId !== referenceColumnId) {
+        }, ETableName.TASK);
+        if (fromColumnId !== referenceColumnId && fromColumnTasks != null && referenceColumnTasks != null) {
             for (const t of fromColumnTasks) {
                 if (t.index > fromTask.index) {
-                    await taskModel.findByIdAndUpdate(t._id, {
+                    await findByIdAndUpdate<ITask>(t._id, {
                         index: t.index - 1
-                    });
+                    }, ETableName.TASK);
                 }
             }
             if (referenceTask != null) {
                 for (const t of referenceColumnTasks) {
                     if (t.index >= referenceTask.index) {
-                        await taskModel.findByIdAndUpdate(t._id, {
+                        await findByIdAndUpdate<ITask>(t._id, {
                             index: t.index + 1
-                        });
+                        }, ETableName.TASK);
                     }
                 }
-                await taskModel.findByIdAndUpdate(fromId, {
+                await findByIdAndUpdate<ITask>(fromId, {
                     columnId: referenceColumnId,
                     index: referenceTask.index
-                });
+                }, ETableName.TASK);
                 return res.status(StatusCodes.OK).json("Task reordered");
             } else {
-                await taskModel.findByIdAndUpdate(fromId, {
+                await findByIdAndUpdate<ITask>(fromId, {
                     columnId: referenceColumnId,
                     index: referenceColumnTasks.length
-                });
+                }, ETableName.TASK);
                 return res.status(StatusCodes.OK).json("Task reordered");
             }
         } else if (
             fromColumnId === referenceColumnId &&
-            referenceTask != null
+            referenceTask != null && referenceColumnTasks != null
         ) {
             if (type === "before") {
                 for (const t of referenceColumnTasks) {
@@ -152,17 +164,17 @@ const reorder = async (
                         t.index > referenceTask.index &&
                         t.index < fromTask.index
                     ) {
-                        await taskModel.findByIdAndUpdate(t._id, {
+                        await findByIdAndUpdate<ITask>(t._id, {
                             index: t.index + 1
-                        });
+                        }, ETableName.TASK);
                     }
                 }
-                await taskModel.findByIdAndUpdate(fromId, {
+                await findByIdAndUpdate<ITask>(fromId, {
                     index: referenceTask.index
-                });
-                await taskModel.findByIdAndUpdate(referenceId, {
+                }, ETableName.TASK);
+                await findByIdAndUpdate<ITask>(referenceId, {
                     index: referenceTask.index + 1
-                });
+                }, ETableName.TASK);
                 return res.status(StatusCodes.OK).json("Task reordered");
             } else if (type === "after") {
                 for (const t of referenceColumnTasks) {
@@ -170,17 +182,17 @@ const reorder = async (
                         t.index > fromTask.index &&
                         t.index < referenceTask.index
                     ) {
-                        await taskModel.findByIdAndUpdate(t._id, {
+                        await findByIdAndUpdate(t._id, {
                             index: t.index - 1
-                        });
+                        }, ETableName.TASK);
                     }
                 }
-                await taskModel.findByIdAndUpdate(referenceId, {
+                await findByIdAndUpdate<ITask>(referenceId, {
                     index: referenceTask.index - 1
-                });
-                await taskModel.findByIdAndUpdate(fromId, {
+                }, ETableName.TASK);
+                await findByIdAndUpdate<ITask>(fromId, {
                     index: referenceTask.index
-                });
+                }, ETableName.TASK);
                 return res.status(StatusCodes.OK).json("Task reordered");
             }
         }
